feat: Add Trixie variant and update CI
This commit introduces support for the Trixie PHP variant in the Docker build process.
It includes the necessary changes to the CI workflow `.github/workflows/docker-buildx.yml` to include `apache-trixie` in the matrix strategy and generate appropriate Docker tags.
Additionally, a new Dockerfile `php8/apache-trixie/Dockerfile` is added to define the build for the Trixie variant.